requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.8.1
    # via
    #   eth-brownie
    #   web3
aiosignal==1.2.0
    # via
    #   aiohttp
    #   eth-brownie
asttokens==2.0.5
    # via
    #   eth-brownie
    #   vyper
async-timeout==4.0.1
    # via
    #   aiohttp
    #   eth-brownie
asynctest==0.13.0
    # via eth-brownie
attrs==21.2.0
    # via
    #   aiohttp
    #   eth-brownie
    #   hypothesis
    #   jsonschema
    #   pytest
base58==2.1.1
    # via
    #   eth-brownie
    #   multiaddr
bitarray==1.2.2
    # via
    #   eth-account
    #   eth-brownie
black==21.11b1
    # via eth-brownie
cached-property==1.5.2
    # via eth-brownie
certifi==2021.10.8
    # via
    #   eth-brownie
    #   requests
charset-normalizer==2.0.9
    # via
    #   aiohttp
    #   eth-brownie
    #   requests
click==8.0.3
    # via
    #   black
    #   eth-brownie
cytoolz==0.11.2
    # via
    #   eth-brownie
    #   eth-keyfile
    #   eth-utils
dataclassy==0.11.1
    # via
    #   eip712
    #   eth-brownie
eip712==0.1.0
    # via eth-brownie
eth-abi==2.1.1
    # via
    #   eip712
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   web3
eth-account==0.5.6
    # via
    #   eth-brownie
    #   web3
eth-brownie==1.17.2
    # via -r requirements.in
eth-event==1.2.3
    # via eth-brownie
eth-hash[pycryptodome]==0.3.2
    # via
    #   eth-brownie
    #   eth-event
    #   eth-utils
    #   web3
eth-keyfile==0.5.1
    # via
    #   eth-account
    #   eth-brownie
eth-keys==0.3.3
    # via
    #   eth-account
    #   eth-brownie
    #   eth-keyfile
eth-rlp==0.2.1
    # via
    #   eth-account
    #   eth-brownie
eth-typing==2.2.2
    # via
    #   eip712
    #   eth-abi
    #   eth-brownie
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==1.10.0
    # via
    #   eip712
    #   eth-abi
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
execnet==1.9.0
    # via
    #   eth-brownie
    #   pytest-xdist
frozenlist==1.2.0
    # via
    #   aiohttp
    #   aiosignal
    #   eth-brownie
hexbytes==0.2.2
    # via
    #   eip712
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   eth-rlp
    #   web3
hypothesis==6.27.3
    # via eth-brownie
idna==3.3
    # via
    #   eth-brownie
    #   requests
    #   yarl
importlib-metadata==4.8.2
    # via eth-brownie
inflection==0.5.0
    # via
    #   eth-brownie
    #   mythx-models
    #   pythx
iniconfig==1.1.1
    # via
    #   eth-brownie
    #   pytest
ipfshttpclient==0.8.0a2
    # via
    #   eth-brownie
    #   web3
jsonschema==3.2.0
    # via
    #   eth-brownie
    #   mythx-models
    #   web3
lazy-object-proxy==1.6.0
    # via eth-brownie
lru-dict==1.1.7
    # via
    #   eth-brownie
    #   web3
multiaddr==0.0.9
    # via
    #   eth-brownie
    #   ipfshttpclient
multidict==5.2.0
    # via
    #   aiohttp
    #   eth-brownie
    #   yarl
mypy-extensions==0.4.3
    # via
    #   black
    #   eth-brownie
mythx-models==1.9.1
    # via
    #   eth-brownie
    #   pythx
netaddr==0.8.0
    # via
    #   eth-brownie
    #   multiaddr
packaging==21.3
    # via
    #   eth-brownie
    #   pytest
parsimonious==0.8.1
    # via
    #   eth-abi
    #   eth-brownie
pathspec==0.9.0
    # via
    #   black
    #   eth-brownie
platformdirs==2.4.0
    # via
    #   black
    #   eth-brownie
pluggy==1.0.0
    # via
    #   eth-brownie
    #   pytest
prompt-toolkit==3.0.23
    # via eth-brownie
protobuf==3.19.1
    # via
    #   eth-brownie
    #   web3
psutil==5.8.0
    # via eth-brownie
py==1.11.0
    # via
    #   eth-brownie
    #   pytest
    #   pytest-forked
py-solc-ast==1.2.9
    # via eth-brownie
py-solc-x==1.1.1
    # via eth-brownie
pycryptodome==3.12.0
    # via
    #   eip712
    #   eth-brownie
    #   eth-hash
    #   eth-keyfile
    #   vyper
pygments==2.10.0
    # via
    #   eth-brownie
    #   pygments-lexer-solidity
pygments-lexer-solidity==0.7.0
    # via eth-brownie
pyjwt==1.7.1
    # via
    #   eth-brownie
    #   pythx
pyparsing==3.0.6
    # via
    #   eth-brownie
    #   packaging
pyrsistent==0.18.0
    # via
    #   eth-brownie
    #   jsonschema
pytest==6.2.5
    # via
    #   eth-brownie
    #   pytest-forked
    #   pytest-xdist
pytest-forked==1.3.0
    # via
    #   eth-brownie
    #   pytest-xdist
pytest-xdist==1.34.0
    # via eth-brownie
python-dateutil==2.8.1
    # via
    #   eth-brownie
    #   mythx-models
    #   pythx
python-dotenv==0.16.0
    # via eth-brownie
pythx==1.6.1
    # via eth-brownie
pyyaml==5.4.1
    # via eth-brownie
regex==2021.11.10
    # via
    #   black
    #   eth-brownie
requests==2.26.0
    # via
    #   eth-brownie
    #   ipfshttpclient
    #   py-solc-x
    #   pythx
    #   vvm
    #   web3
rlp==2.0.1
    # via
    #   eth-account
    #   eth-brownie
    #   eth-rlp
semantic-version==2.8.5
    # via
    #   eth-brownie
    #   py-solc-x
    #   vvm
    #   vyper
six==1.16.0
    # via
    #   asttokens
    #   eth-brownie
    #   jsonschema
    #   multiaddr
    #   parsimonious
    #   pytest-xdist
    #   python-dateutil
sortedcontainers==2.4.0
    # via
    #   eth-brownie
    #   hypothesis
toml==0.10.2
    # via
    #   eth-brownie
    #   pytest
tomli==1.2.2
    # via
    #   black
    #   eth-brownie
toolz==0.11.2
    # via
    #   cytoolz
    #   eth-brownie
tqdm==4.62.3
    # via eth-brownie
typed-ast==1.5.1
    # via eth-brownie
typing-extensions==3.10.0.2
    # via
    #   async-timeout
    #   black
    #   eth-brownie
urllib3==1.26.7
    # via
    #   eth-brownie
    #   requests
varint==1.0.2
    # via
    #   eth-brownie
    #   multiaddr
vvm==0.1.0
    # via eth-brownie
vyper==0.3.1
    # via eth-brownie
wcwidth==0.2.5
    # via
    #   eth-brownie
    #   prompt-toolkit
web3==5.25.0
    # via eth-brownie
websockets==9.1
    # via
    #   eth-brownie
    #   web3
wrapt==1.13.3
    # via eth-brownie
yarl==1.7.2
    # via
    #   aiohttp
    #   eth-brownie
zipp==3.6.0
    # via
    #   eth-brownie
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
